+++ test/fetch.spec.js
@@ -6,12 +6,13 @@ import {Action,Store} from '../src/fizz';
 
 /* eslint-disable no-magic-numbers */
 
-test.skip('fizz: fetch', function(t) {
+test.only('fizz: fetch', function(t) {
   // Will be skipped for the moment. The test used the web api from fizz-
   // example reverse and will fail if reverse is not running.
   let reverse = Action.create(function(str: string) {
     return fetch(
-      '//localhost:8000/reverse/' + encodeURI(str)
+      '//localhost:8000/reverse/' + encodeURI(str),
+      {mode: 'no-cors'}
     );
   });