+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.1.6 / 2016-09-01
 ------------------
 - Processing of fetch promises.
+- Using isPromise from jsz-isTyp insteed of instanceof.
 
 0.1.5 / 2016-08-30
 ------------------